Support turn based anime game version 2.4.0

+++ Common/Data/Excel/ShopGoodsConfigExcel.cs
@@ -35,7 +35,7 @@ public class ShopGoodsConfigExcel : ExcelResource
     {
         return new Goods
         {
-            EndTime = long.MaxValue,
+            EndTime = uint.MaxValue,
             GoodsId = (uint)GoodsID,
             ItemId = (uint)ItemID
         };
